// common/types/user.types.d.ts
export declare enum UserRole {
    ADMIN = "ADMIN",
    READER = "READER"
}
export interface User {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    role: UserRole;
    createdAt: string;
    updatedAt: string;
}
export interface CreateUserRequest {
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    role: UserRole;
}
export interface UpdateUserRequest {
    email?: string;
    firstName?: string;
    lastName?: string;
    role?: UserRole;
}
export interface LoginRequest {
    email: string;
    password: string;
}
export interface AuthResponse {
    user: User;
    accessToken: string;
}
//# sourceMappingURL=user.types.d.ts.map